+++ examples/with-ssg/src/pages/layout.tsx
@@ -1,0 +1,12 @@
+import { Outlet } from 'ice';
+
+export default function Layout() {
+  console.log('render layout');
+
+  return (
+    <>
+      <h2>With SSG</h2>
+      <Outlet />
+    </>
+  );
+}