+++ packages/ice/src/commands/build.ts
@@ -1,5 +1,4 @@
 import * as path from 'path';
-import consola from 'consola';
 import fse from 'fs-extra';
 import { getWebpackConfig } from '@ice/webpack-config';
 import type { Context, TaskConfig } from 'build-scripts';
@@ -8,13 +7,15 @@ import type { StatsError, Stats } from 'webpack';
 import type { Config } from '@ice/webpack-config/esm/types';
 import type ora from '@ice/bundles/compiled/ora/index.js';
 import type { AppConfig } from '@ice/runtime/esm/types';
-import type { ServerCompiler, GetAppConfig, GetRoutesConfig, ExtendsPluginAPI } from '../types/plugin.js';
+import type { ServerCompiler, GetAppConfig, GetRoutesConfig, ExtendsPluginAPI, GetDataloaderConfig } from '../types/plugin.js';
 import webpackCompiler from '../service/webpackCompiler.js';
 import formatWebpackMessages from '../utils/formatWebpackMessages.js';
-import { RUNTIME_TMP_DIR, SERVER_OUTPUT_DIR } from '../constant.js';
+import { IMPORT_META_RENDERER, IMPORT_META_TARGET, RUNTIME_TMP_DIR, SERVER_OUTPUT_DIR } from '../constant.js';
 import emptyDir from '../utils/emptyDir.js';
 import type { UserConfig } from '../types/userConfig.js';
 import warnOnHashRouterEnabled from '../utils/warnOnHashRouterEnabled.js';
+import { logger } from '../utils/logger.js';
+import { getExpandedEnvs } from '../utils/runtimeEnv.js';
 
 const build = async (
   context: Context<Config, ExtendsPluginAPI>,
@@ -25,6 +26,7 @@ const build = async (
     getAppConfig: GetAppConfig;
     appConfig: AppConfig;
     getRoutesConfig: GetRoutesConfig;
+    getDataloaderConfig: GetDataloaderConfig;
     userConfigHash: string;
     userConfig: UserConfig;
   },
@@ -36,10 +38,12 @@ const build = async (
     getAppConfig,
     appConfig,
     getRoutesConfig,
+    getDataloaderConfig,
     userConfigHash,
     userConfig,
   } = options;
-  const { applyHook, rootDir } = context;
+  const { applyHook, commandArgs, rootDir } = context;
+  const { target } = commandArgs;
 
   if (appConfig?.router?.type === 'hash') {
     warnOnHashRouterEnabled(userConfig);
@@ -52,6 +56,11 @@ const build = async (
     webpack,
     runtimeTmpDir: RUNTIME_TMP_DIR,
     userConfigHash,
+    getExpandedEnvs,
+    runtimeDefineVars: {
+      [IMPORT_META_TARGET]: JSON.stringify(target),
+      [IMPORT_META_RENDERER]: JSON.stringify('client'),
+    },
   }));
   const outputDir = webpackConfigs[0].output.path;
 
@@ -60,6 +69,7 @@ const build = async (
     serverCompiler,
     getAppConfig,
     getRoutesConfig,
+    getDataloaderConfig,
   };
   const compiler = await webpackCompiler({
     context,
@@ -96,7 +106,7 @@ const build = async (
       }
 
       if (messages.errors.length) {
-        consola.error('webpack compile error');
+        logger.error('Webpack compile error.');
         reject(new Error(messages.errors.join('\n\n')));
         return;
       } else {
@@ -122,6 +132,7 @@ const build = async (
     output,
     getAppConfig,
     getRoutesConfig,
+    getDataloaderConfig,
     appConfig,
   });