+++ packages/ice/src/plugins/web/index.ts
@@ -1,5 +1,4 @@
 import * as path from 'path';
-import consola from 'consola';
 import chalk from 'chalk';
 import type { RenderMode } from '@ice/runtime';
 import lodash from '@ice/bundles/compiled/lodash/index.js';
@@ -7,12 +6,13 @@ import type { Plugin } from '../../types/plugin.js';
 import ReCompilePlugin from '../../webpack/ReCompilePlugin.js';
 import DataLoaderPlugin from '../../webpack/DataLoaderPlugin.js';
 import { getRouteExportConfig } from '../../service/config.js';
-import { WEB, SERVER_OUTPUT_DIR } from '../../constant.js';
+import { WEB, SERVER_OUTPUT_DIR, IMPORT_META_TARGET, IMPORT_META_RENDERER } from '../../constant.js';
 import getWebTask from '../../tasks/web/index.js';
 import generateEntry from '../../utils/generateEntry.js';
 import openBrowser from '../../utils/openBrowser.js';
 import getServerCompilerPlugin from '../../utils/getServerCompilerPlugin.js';
 import type ServerCompilerPlugin from '../../webpack/ServerCompilerPlugin.js';
+import { logger } from '../../utils/logger.js';
 
 const { debounce } = lodash;
 
@@ -20,7 +20,8 @@ const plugin: Plugin = () => ({
   name: 'plugin-web',
   setup: ({ registerTask, onHook, context, generator, serverCompileTask, dataCache, watch, getAllPlugin }) => {
     const { rootDir, commandArgs, command, userConfig } = context;
-    registerTask(WEB, getWebTask({ rootDir, command, dataCache }));
+
+    registerTask(WEB, getWebTask({ rootDir, command, dataCache, userConfig }));
 
     generator.addExport({
       specifier: ['Link', 'Outlet', 'useParams', 'useSearchParams', 'useLocation', 'useNavigate'],
@@ -65,6 +66,11 @@ const plugin: Plugin = () => ({
         serverCompileTask: command === 'start' ? serverCompileTask : null,
         userConfig,
         ensureRoutesConfig,
+        runtimeDefineVars: {
+          [IMPORT_META_TARGET]: JSON.stringify('web'),
+          [IMPORT_META_RENDERER]: JSON.stringify('server'),
+        },
+        incremental: command === 'start',
       });
       webpackConfigs[0].plugins.push(
         // Add webpack plugin of data-loader in web task
@@ -86,10 +92,8 @@ const plugin: Plugin = () => ({
           }),
         );
         const debounceCompile = debounce(() => {
+          serverCompilerPlugin?.buildResult?.rebuild();
           console.log('Document updated, try to reload page for latest html content.');
-          if (serverCompilerPlugin) {
-            serverCompilerPlugin.compileTask();
-          }
         }, 200);
         watch.addEvent([
           /src\/document(\/index)?(.js|.jsx|.tsx)/,
@@ -134,21 +138,21 @@ const plugin: Plugin = () => ({
 
     onHook('after.start.compile', async ({ isSuccessful, isFirstCompile, urls, devUrlInfo }) => {
       const { port, open } = commandArgs;
-      const { devPath, hashChar } = devUrlInfo;
+      const { devPath } = devUrlInfo;
       if (isSuccessful && isFirstCompile) {
         let logoutMessage = '\n';
         logoutMessage += chalk.green(' Starting the development server at:');
         if (process.env.CLOUDIDE_ENV) {
-          logoutMessage += `\n   - IDE server: https://${process.env.WORKSPACE_UUID}-${port}.${process.env.WORKSPACE_HOST}${hashChar}${devPath}`;
+          logoutMessage += `\n   - IDE server: https://${process.env.WORKSPACE_UUID}-${port}.${process.env.WORKSPACE_HOST}${devPath}`;
         } else {
           logoutMessage += `\n
-    - Local  : ${chalk.underline.white(`${urls.localUrlForBrowser}${hashChar}${devPath}`)}
-    - Network: ${chalk.underline.white(`${urls.lanUrlForTerminal}${hashChar}${devPath}`)}`;
+    - Local  : ${chalk.underline.white(`${urls.localUrlForBrowser}${devPath}`)}
+    - Network: ${chalk.underline.white(`${urls.lanUrlForTerminal}${devPath}`)}`;
         }
-        consola.log(`${logoutMessage}\n`);
+        logger.log(`${logoutMessage}\n`);
 
         if (open) {
-          openBrowser(`${urls.localUrlForBrowser}${hashChar}${devPath}`);
+          openBrowser(`${urls.localUrlForBrowser}${devPath}`);
         }
       }
     });