+++ packages/ice/src/routes.ts
@@ -35,7 +35,7 @@ export async function generateRoutesInfo(rootDir: string, routesConfig: UserConf
     routes,
     loaders: generateRouteConfig(routes, 'dataLoader', (str, imports) => {
       return imports.length > 0 ? `${str}
-const loaders = {
+export default {
   ${imports.map(([routeId, importKey]) => `'${routeId}': ${importKey},`).join('\n  ')}
 };` : '';
     }),