+++ packages/ice/src/service/webpackCompiler.ts
@@ -1,11 +1,11 @@
 import webpackBundler from '@ice/bundles/compiled/webpack/index.js';
 import type ora from '@ice/bundles/compiled/ora/index.js';
-import consola from 'consola';
 import type { TaskConfig, Context } from 'build-scripts';
 import type { Config } from '@ice/webpack-config/esm/types';
 import type webpack from 'webpack';
-import type { Urls, ServerCompiler, GetAppConfig, GetRoutesConfig, ExtendsPluginAPI } from '../types/plugin.js';
+import type { Urls, ServerCompiler, GetAppConfig, GetRoutesConfig, ExtendsPluginAPI, GetDataloaderConfig } from '../types/plugin.js';
 import formatWebpackMessages from '../utils/formatWebpackMessages.js';
+import { logger } from '../utils/logger.js';
 
 async function webpackCompiler(options: {
   context: Context<Config, ExtendsPluginAPI>;
@@ -18,6 +18,7 @@ async function webpackCompiler(options: {
     serverCompiler: ServerCompiler;
     getAppConfig: GetAppConfig;
     getRoutesConfig: GetRoutesConfig;
+    getDataloaderConfig: GetDataloaderConfig;
   };
 }) {
   const {
@@ -52,8 +53,8 @@ async function webpackCompiler(options: {
     // @ts-ignore
     compiler = webpackBundler(webpackConfigs);
   } catch (err) {
-    consola.error('Webpack compile error.');
-    consola.error(err.message || err);
+    logger.error('Webpack compile error.');
+    logger.error(err.message || err);
   }
 
   let isFirstCompile = true;
@@ -74,16 +75,14 @@ async function webpackCompiler(options: {
       if (messages.errors.length > 1) {
         messages.errors.length = 1;
       }
-      consola.error('Compiled with errors.');
+      logger.error('Compiled with errors.');
       console.error(messages.errors.join('\n'));
       return;
     } else if (messages.warnings.length) {
-      consola.warn('Compiled with warnings.');
-      consola.warn(messages.warnings.join('\n'));
+      logger.warn('Compiled with warnings.');
+      logger.warn(messages.warnings.join('\n'));
     }
     if (command === 'start') {
-      const appConfig = (await hooksAPI.getAppConfig()).default;
-      const hashChar = appConfig?.router?.type === 'hash' ? '#/' : '';
       // compiler.hooks.done is AsyncSeriesHook which does not support async function
       await applyHook('after.start.compile', {
         stats,
@@ -91,7 +90,6 @@ async function webpackCompiler(options: {
         isFirstCompile,
         urls,
         devUrlInfo: {
-          hashChar,
           devPath,
         },
         messages,