+++ packages/ice/src/types/plugin.ts
@@ -1,18 +1,19 @@
 import type webpack from '@ice/bundles/compiled/webpack';
 import type { _Plugin, CommandArgs, TaskConfig } from 'build-scripts';
 import type { Configuration, Stats, WebpackOptionsNormalized } from '@ice/bundles/compiled/webpack';
-import type { BuildOptions, BuildResult } from 'esbuild';
+import type { esbuild } from '@ice/bundles';
 import type { NestedRouteManifest } from '@ice/route-manifest';
 import type { Config } from '@ice/webpack-config/esm/types';
 import type { AppConfig, AssetsManifest } from '@ice/runtime/esm/types';
 import type { DeclarationData, AddRenderFile, AddTemplateFiles, ModifyRenderData, AddDataLoaderImport, Render } from './generator.js';
 
 type AddExport = (exportData: DeclarationData) => void;
+type AddEntryCode = (callback: (code: string) => string) => void;
 type RemoveExport = (removeSource: string | string[]) => void;
 type EventName = 'add' | 'addDir' | 'change' | 'unlink' | 'unlinkDir';
 
 type ServerCompilerBuildOptions = Pick<
-  BuildOptions,
+  esbuild.BuildOptions,
   'write' |
   'target' |
   'minify' |
@@ -29,8 +30,11 @@ type ServerCompilerBuildOptions = Pick<
   'plugins' |
   'logLevel' |
   'sourcemap' |
-  'metafile'
+  'metafile' |
+  'incremental'
 >;
+
+export type ServerBuildResult = Partial<esbuild.BuildResult & { serverEntry: string; error: any }>;
 
 export type ServerCompiler = (
   buildOptions: ServerCompilerBuildOptions,
@@ -39,11 +43,15 @@ export type ServerCompiler = (
     preBundle?: boolean;
     externalDependencies?: boolean;
     transformEnv?: boolean;
-    assetsManifest?: AssetsManifest;
+    compilationInfo?: {
+      assetsManifest?: AssetsManifest;
+    };
     redirectImports?: Config['redirectImports'];
     removeOutputs?: boolean;
+    runtimeDefineVars?: Record<string, string>;
+    enableEnv?: boolean;
   }
-) => Promise<Partial<BuildResult & { serverEntry: string; error: any }>>;
+) => Promise<ServerBuildResult>;
 export type WatchEvent = [
   pattern: RegExp | string,
   event: (eventName: EventName, filePath: string) => void,
@@ -59,6 +67,7 @@ export interface Urls {
 
 export type GetAppConfig = (exportNames?: string[]) => Promise<any>;
 export type GetRoutesConfig = (specifyRoutId?: string) => Promise<any>;
+export type GetDataloaderConfig = (specifyRoutId?: string) => Promise<any>;
 
 interface BeforeCommandRunOptions {
   commandArgs: CommandArgs;
@@ -67,6 +76,7 @@ interface BeforeCommandRunOptions {
   urls?: Urls;
   getAppConfig: GetAppConfig;
   getRoutesConfig: GetRoutesConfig;
+  getDataloaderConfig: GetDataloaderConfig;
   serverCompiler: ServerCompiler;
 }
 
@@ -88,7 +98,6 @@ interface AfterCommandCompileOptions {
 
 interface DevServerInfo {
   devPath: string;
-  hashChar: string;
 }
 
 export interface HookLifecycle {
@@ -122,6 +131,7 @@ export interface ExtendsPluginAPI {
     modifyRenderData: ModifyRenderData;
     render: Render;
     addDataLoaderImport: AddDataLoaderImport;
+    addEntryCode: AddEntryCode;
   };
   watch: {
     addEvent?: (watchEvent: WatchEvent) => void;