+++ packages/plugin-pha/src/generateManifest.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 import * as fs from 'fs';
-import type { GetAppConfig, GetRoutesConfig, ServerCompiler } from '@ice/app/esm/types';
+import type { GetAppConfig, GetDataloaderConfig, GetRoutesConfig, ServerCompiler } from '@ice/app/esm/types';
 import { parseManifest, rewriteAppWorker, getAppWorkerUrl, getMultipleManifest, type ParseOptions } from './manifestHelpers.js';
 import type { Compiler } from './index.js';
 
@@ -11,6 +11,7 @@ export interface Options {
   compiler: Compiler;
   getAppConfig: GetAppConfig;
   getRoutesConfig: GetRoutesConfig;
+  getDataloaderConfig: GetDataloaderConfig;
   compileTask?: () => ReturnType<ServerCompiler>;
 }
 
@@ -37,9 +38,10 @@ export default async function generateManifest({
   parseOptions,
   getAppConfig,
   getRoutesConfig,
+  getDataloaderConfig,
   compiler,
 }: Options) {
-  const [appConfig, routesConfig] = await Promise.all([getAppConfig(['phaManifest']), getRoutesConfig()]);
+  const [appConfig, routesConfig, dataloaderConfig] = await Promise.all([getAppConfig(['phaManifest']), getRoutesConfig(), getDataloaderConfig()]);
   let manifest = appConfig.phaManifest;
   const appWorkerPath = getAppWorkerUrl(manifest, path.join(rootDir, 'src'));
   if (appWorkerPath) {
@@ -51,6 +53,7 @@ export default async function generateManifest({
     });
   }
   const phaManifest = await parseManifest(manifest, {
+    dataloaderConfig,
     ...parseOptions,
     routesConfig,
   } as ParseOptions);