+++ packages/runtime/src/runServerApp.tsx
@@ -203,6 +203,7 @@ async function doRender(serverContext: ServerContext, renderOptions: RenderOptio
     appData,
     routesData: null,
     routesConfig: null,
+    renderMode,
     assetsManifest,
     basename: finalBasename,
     matches: [],
@@ -228,9 +229,7 @@ async function doRender(serverContext: ServerContext, renderOptions: RenderOptio
   }
 
   const matches = matchRoutes(routes, location, finalBasename);
-  if (!matches.length) {
-    return render404();
-  }
+
 
   if (distType === 'javascript' || (Array.isArray(distType) && distType.includes('javascript'))
   ) {
@@ -241,6 +240,8 @@ async function doRender(serverContext: ServerContext, renderOptions: RenderOptio
 
   if (documentOnly) {
     return renderDocument({ matches, routePath, renderOptions });
+  } else if (!matches.length) {
+    return render404();
   }
   try {
     const routeModules = await loadRouteModules(matches.map(({ route: { id, load } }) => ({ id, load })));