+++ packages/runtime/src/types.ts
@@ -1,7 +1,7 @@
 import type { IncomingMessage, ServerResponse } from 'http';
 import type { Action, InitialEntry, Location } from 'history';
 import type { ComponentType, ReactNode, PropsWithChildren } from 'react';
-import type { HydrationOptions } from 'react-dom/client';
+import type { HydrationOptions, Root } from 'react-dom/client';
 import type { Navigator, Params } from 'react-router-dom';
 
 type UseConfig = () => RouteConfig<Record<string, any>>;
@@ -37,7 +37,7 @@ export interface AppExport {
 export type DataLoaderResult = (Promise<RouteData> | RouteData) | RouteData;
 export type DataLoader = (ctx: RequestContext) => DataLoaderResult;
 
-interface StaticDataLoader {
+export interface StaticDataLoader {
   key?: string;
   prefetch_type?: string;
   api: string;
@@ -87,18 +87,19 @@ export interface AppContext {
   appExport?: AppExport;
   basename?: string;
   downgrade?: boolean;
+  renderMode?: string;
 }
 
 export type WindowContext = Pick<
   AppContext,
-  'appData' | 'routesData' | 'routesConfig' | 'routePath' | 'downgrade' | 'matchedIds' | 'documentOnly'
+  'appData' | 'routesData' | 'routesConfig' | 'routePath' | 'downgrade' | 'matchedIds' | 'documentOnly' | 'renderMode'
 >;
 
 export type Renderer = (
   container: Element | Document,
   initialChildren: React.ReactNode,
   options?: HydrationOptions,
-) => void;
+) => Root;
 
 export interface ServerContext {
   req?: IncomingMessage;
@@ -147,6 +148,7 @@ export type AppProvider = ComponentWithChildren<any>;
 export type RouteWrapper = ComponentType<any>;
 
 export type SetAppRouter = (AppRouter: ComponentType<AppRouterProps>) => void;
+export type GetAppRouter = () => AppProvider;
 export type AddProvider = (Provider: AppProvider) => void;
 export type SetRender = (render: Renderer) => void;
 export type AddWrapper = (wrapper: RouteWrapper, forLayout?: boolean) => void;
@@ -171,6 +173,7 @@ export interface AssetsManifest {
 
 export interface RuntimeAPI {
   setAppRouter?: SetAppRouter;
+  getAppRouter: GetAppRouter;
   addProvider: AddProvider;
   setRender: SetRender;
   addWrapper: AddWrapper;
@@ -186,26 +189,26 @@ export interface StaticRuntimeAPI {
   };
 }
 
-export interface RuntimePlugin {
+export interface RuntimePlugin<T = Record<string, any>> {
   (
     apis: RuntimeAPI,
-    runtimeOptions?: Record<string, any>,
+    runtimeOptions?: T,
   ): Promise<void> | void;
 }
 
-export interface RuntimePlugin {
+export interface StaticRuntimePlugin<T = Record<string, any>> {
   (
     apis: StaticRuntimeAPI,
-    runtimeOptions?: Record<string, any>,
+    runtimeOptions?: T,
   ): Promise<void> | void;
 }
 
 export interface CommonJsRuntime {
-  default: RuntimePlugin;
+  default: RuntimePlugin | StaticRuntimePlugin;
 }
 
 export interface RuntimeModules {
-  statics?: (RuntimePlugin | CommonJsRuntime)[];
+  statics?: (StaticRuntimePlugin | CommonJsRuntime)[];
   commons?: (RuntimePlugin | CommonJsRuntime)[];
 }