+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ice/runtime
 
+## v1.0.5
+
+- [feat] support template parse for dataLoader
+
+## v1.0.4
+
+- [chore] show more log when failed to load route module
+- [fix] rerender after hydrate for ssg page
+
+## v1.0.3
+
+- [feat] optimize log when load route module failed
+- [revert] window context merge order
+- [fix] should call dataLoader when hydrate for SSG
+- [fix] pass dataLoader as list
+
 ## v1.0.2
 
 - [fix] window context merge order